Second attempt at refactoring parser.
14 test fail (I think mostly due to MariaDB syntax issues).
